Add a function to navigate the json buffer (#20)
* Add `jsonian-enclosing-item`

* Complete `parse-path` & `find-siblings`

These are subsideary functions necessary to implementing the programatic
comletion necessary for `jsonian-find`.

* Add valid-path-p

* All functions added, they don't work yet

* Basic functionality now working

- The next step is speeding up completion, which I think can be done
with caching. We can use `before-change-hook` to ensure buffer-local
cached values are accurate.
- Emacs expects that if we select an object, the next `[` or `.` is
inserted as part of the completion. We don't do that yet.
- We don't start the completion when given the next `.`. We wait until
the next character is typed.

* Add caching

* Add completing-sort

This upps the required emacs version to 27.1

* Fix lint

* Fix bug in caching

* Prevent recursion on leaf types

* Prevent overwriting for leaf types

* Fix lint

* Parse partial paths

* Improve tests

* Handle ending open [ when parsing paths

* Cleanup partial parsing of input properties

* Add the ability to display cached types

* Add alignment and type previews

* Handle top level paths

This is to make path mandatory and thus avoid the previous nil check.
This was necessary since an empty path (top level item) is also nil.

* Display the opening `[` for array indexing

* Add properties

* Organize jsonian.el

* Fix lint

* Fix other lints