Remove `'object` from `jsonian--path` outputs
The `'object` marker was used to facilitate collecting the path from
objects such as:

```json
{
  "a": null,
  "b":█true
}
```

It was necessary because at the time, both the "a" and "b" field tags
were collected as part of `jsonian--path` and I needed a way to
distinguish which tags were worth collecting. During post-processing in
`jsonian--reconstruct-path`, the first collected field would be kept,
while the rest were discarded until a non-field path element was
detected.

`'object` was necessary to distinguish between the paths from the
following blobs:

```json
{ "a": 0, "b": 0, "c": 0█ }
```

```json
{ "a": 0, "b": { "c": 0█ } }
```